convert date to local date, and setup ktor client

+++ app/src/main/java/dev/sobhy/healthhubfordoctors/authfeature/presentation/register/RegisterViewModel.kt
@@ -54,7 +54,7 @@ class RegisterViewModel
                     email = registerState.value.email,
                     password = registerState.value.password,
                     phone = registerState.value.phone,
-                    dateOfBirth = registerState.value.dateOfBirth.toLong(),
+                    dateOfBirth = registerState.value.dateOfBirth,
                     gender = registerState.value.gender,
                     professionalTitle = registerState.value.professionalTitle,
                     specialization = registerState.value.specialization,
